Move out the reinterpret stuff to ReinterpretFramebuffer.cpp. Draw2D refactor.
