+++ GPU/Common/FramebufferManagerCommon.h
@@ -361,6 +361,7 @@ protected:
 	void DrawActiveTexture(float x, float y, float w, float h, float destW, float destH, float u0, float v0, float u1, float v1, int uvRotation, int flags);
 
 	void DrawStrip2D(Draw::Texture *tex, Draw2DVertex *verts, int vertexCount, bool linearFilter, RasterChannel channel);
+	void Ensure2DResources();
 
 	bool UpdateSize();
